requirements.txt
fastapi==0.114.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
SQLAlchemy==2.0.35
pydantic==2.9.1
python-dotenv==1.0.1